Edded new graphs
Bw. requested a network map with 3 classes (lead authors from Africa, lead authors from overseas, and lead authors affiliated both to Africa and overseas).
a copy of the code was made and edited to that purpose. The new graphs can also be found in the plot folder, with a new folder having figma edits to add a comprehensive legend, with or without the country counts cercles.